# workspace/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_StoreChangeProposalApprovalResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoreChangeProposalApprovalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiResponseStatus"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "success",
          "status",
          "data",
          "error",
          "timestamp"
        ],
        "title": "ApiEnvelope[StoreChangeProposalApprovalResponse]",
        "type": "object"
      },
      "ApiEnvelope_StoreChangeProposalResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoreChangeProposalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiResponseStatus"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "success",
          "status",
          "data",
          "error",
          "timestamp"
        ],
        "title": "ApiEnvelope[StoreChangeProposalResponse]",
        "type": "object"
      },
...
      "ApiEnvelope_SyncJobStatusResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncJobStatusResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiResponseStatus"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "success",
          "status",
          "data",
          "error",
          "timestamp"
        ],
        "title": "ApiEnvelope[SyncJobStatusResponse]",
        "type": "object"
      },
      "ApiError": {
        "additionalProperties": false,
        "description": "Safe error body returned at the MapKeeper API boundary.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "CreateStoreChangeProposalRequest": {
        "additionalProperties": false,
        "description": "Recognized text submitted to create a store change proposal.",
        "properties": {
          "locale": {
            "const": "ko-KR",
            "title": "Locale",
            "type": "string"
          },
          "recognizedText": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Recognizedtext",
            "type": "string"
          },
          "storeProfileId": {
            "format": "uuid",
            "title": "Storeprofileid",
            "type": "string"
          }
        },
        "required": [
          "storeProfileId",
          "recognizedText",
          "locale"
        ],
        "title": "CreateStoreChangeProposalRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Failure envelope: ``success`` is false, ``data`` is null and ``error`` is set.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiResponseStatus"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "success",
          "status",
          "data",
          "error",
          "timestamp"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PatchStoreChangeProposalRequest": {
        "additionalProperties": false,
        "description": "Validated proposal changes replacing the current draft changes.",
        "properties": {
          "changes": {
            "items": {
              "discriminator": {
                "mapping": {
                  "businessHours": "#/components/schemas/BusinessHoursChange",
                  "parkingInfo": "#/components/schemas/ParkingInfoChange",
                  "representativeMenuName": "#/components/schemas/RepresentativeMenuNameChange",
                  "temporaryClosure": "#/components/schemas/TemporaryClosureChange"
                },
                "propertyName": "field"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BusinessHoursChange"
                },
                {
                  "$ref": "#/components/schemas/TemporaryClosureChange"
                },
                {
                  "$ref": "#/components/schemas/RepresentativeMenuNameChange"
                },
                {
                  "$ref": "#/components/schemas/ParkingInfoChange"
                }
              ]
            },
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          }
        },
        "required": [
          "changes"
        ],
        "title": "PatchStoreChangeProposalRequest",
        "type": "object"
      },
...
      "StoreChangeProposalResponse": {
        "additionalProperties": false,
        "description": "Store change proposal returned for review.",
        "properties": {
          "changes": {
            "items": {
              "discriminator": {
                "mapping": {
                  "businessHours": "#/components/schemas/BusinessHoursChange",
                  "parkingInfo": "#/components/schemas/ParkingInfoChange",
                  "representativeMenuName": "#/components/schemas/RepresentativeMenuNameChange",
                  "temporaryClosure": "#/components/schemas/TemporaryClosureChange"
                },
                "propertyName": "field"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BusinessHoursChange"
                },
                {
                  "$ref": "#/components/schemas/TemporaryClosureChange"
                },
                {
                  "$ref": "#/components/schemas/RepresentativeMenuNameChange"
                },
                {
                  "$ref": "#/components/schemas/ParkingInfoChange"
                }
              ]
            },
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          },
          "proposalId": {
            "format": "uuid",
            "title": "Proposalid",
            "type": "string"
          },
          "recognizedTextMasked": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Recognizedtextmasked",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "unmappedRequests": {
            "default": [],
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Unmappedrequests",
            "type": "array"
          }
        },
        "required": [
          "proposalId",
          "recognizedTextMasked",
          "changes",
          "status"
        ],
        "title": "StoreChangeProposalResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "MapKeeper MVP v0.2. JSON fields are camelCase, ids are UUID strings and timestamps are UTC ISO 8601. Undefined request and response fields are rejected.",
    "title": "MapKeeper API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/store-change-proposals": {
      "post": {
        "description": "Turn recognized speech into a masked, structured DRAFT proposal.",
        "operationId": "create_proposal_api_v1_store_change_proposals_post",
        "parameters": [
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreChangeProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_StoreChangeProposalResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a store change proposal",
        "tags": [
          "store-change"
        ]
      }
    },
    "/api/v1/store-change-proposals/{proposalId}": {
      "patch": {
        "description": "Replace the full change list after stale currentValue checking.",
        "operationId": "patch_proposal_api_v1_store_change_proposals__proposalId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposalid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchStoreChangeProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_StoreChangeProposalResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace the changes of a DRAFT proposal",
        "tags": [
          "store-change"
        ]
      }
    },
    "/api/v1/store-change-proposals/{proposalId}/approve": {
      "post": {
        "description": "Commit approval and queue synchronization only after the commit succeeds.",
        "operationId": "approve_proposal_api_v1_store_change_proposals__proposalId__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposalid",
              "type": "string"
            }
          },
          {
            "description": "Required on every approval. Same actor and key with the same approval target replays the existing SyncJob; a different target is 409 IDEMPOTENCY_CONFLICT.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Required on every approval. Same actor and key with the same approval target replays the existing SyncJob; a different target is 409 IDEMPOTENCY_CONFLICT.",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_StoreChangeProposalApprovalResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve a DRAFT proposal and start synchronization",
        "tags": [
          "store-change"
        ]
      }
    },
    "/api/v1/store-change-proposals/{proposalId}/reject": {
      "post": {
        "description": "Move a DRAFT proposal to REJECTED. The request has no body.",
        "operationId": "reject_proposal_api_v1_store_change_proposals__proposalId__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposalid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_StoreChangeProposalResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested transition cannot be performed in the current state. error.code is one of: INVALID_STATE, STALE_PROPOSAL, IDEMPOTENCY_CONFLICT, NO_RETRYABLE_TASKS."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject a DRAFT proposal",
        "tags": [
          "store-change"
        ]
      }
    },
...
    "/api/v1/sync-jobs/{syncJobId}": {
      "get": {
        "description": "Report job progress. A failed platform still returns 200 with status SUCCESS.",
        "operationId": "get_sync_job_api_v1_sync_jobs__syncJobId__get",
        "parameters": [
          {
            "in": "path",
            "name": "syncJobId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Syncjobid",
              "type": "string"
            }
          },
          {
            "description": "Optional client trace id. The server generates one when it is absent.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional client trace id. The server generates one when it is absent.",
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SyncJobStatusResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unparseable JSON or an unusable request combination. error.code is one of: MALFORMED_REQUEST."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Firebase ID Token is missing or invalid. error.code is one of: AUTHENTICATION_REQUIRED."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No resource matches the requested id. error.code is one of: RESOURCE_NOT_FOUND."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Header, path or body field validation failed. error.code is one of: VALIDATION_ERROR."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "MapKeeper API request limit reached. error.code is one of: REQUEST_RATE_LIMITED."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal error replaced with a safe message. error.code is one of: INTERNAL_SERVER_ERROR."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read the aggregate job status and per-platform detail",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}